requirements.txt
google-cloud-bigquery==3.25.0
google-cloud-storage==2.18.2
google-auth==2.35.0
google-api-core==2.20.0
requests==2.32.3
pandas==2.2.2